requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
ddgs>=9.0.0
duckduckgo-search>=5.0.0
loguru>=0.7.2
pydantic>=2.6.0
google-api-python-client>=2.165.0
google-auth>=2.38.0
google-auth-httplib2>=0.2.0
